# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "ShuttlePub development API", "version": "0.1.0", "description": "Unauthenticated development API. In-memory persistence, eventual timeline projection. No Mastodon or Misskey compatibility."},
  "paths": {
    "/api/notes": {"post": {"operationId": "createNote", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateNote"}}}}, "responses": {"201": {"description": "Persisted; projection is eventual"}, "409": {"description": "Duplicate or domain rejection"}, "422": {"description": "Invalid JSON representation"}}}},
    "/api/notes/{id}/reactions": {"post": {"operationId": "react", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["actor", "reaction"], "properties": {"actor": {"type": "string", "format": "uuid"}, "reaction": {"type": "string", "minLength": 1}}}}}}, "responses": {"204": {"description": "Reaction set; repeating the same value is a no-op"}, "409": {"description": "Missing note or invalid reaction"}}}},
    "/api/timeline": {"get": {"operationId": "timeline", "parameters": [{"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 20}}, {"name": "before", "in": "query", "schema": {"type": "integer", "minimum": 0}}], "responses": {"200": {"description": "Newest creation first; reaction changes do not reorder notes", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/TimelineEntry"}}}}}, "400": {"description": "Invalid pagination"}}}},
    "/inbox": {"post": {"operationId": "followInbox", "description": "RSA-SHA256 draft HTTP Signature required over (request-target), host, date, digest, content-type. Date tolerance 300 seconds. Only configured remote origins are allowed. Only Follow is supported.", "requestBody": {"required": true, "content": {"application/activity+json": {"schema": {"type": "object", "required": ["id", "type", "actor", "object"], "properties": {"id": {"type": "string", "format": "uri"}, "type": {"const": "Follow"}, "actor": {"type": "string", "format": "uri"}, "object": {"type": "string", "format": "uri"}}}}}}, "responses": {"202": {"description": "Accept delivered successfully"}, "400": {"description": "Unsupported activity or malformed body"}, "401": {"description": "Invalid signature or identity"}, "403": {"description": "Remote origin not allowed"}, "413": {"description": "Body exceeds 256 KiB"}, "502": {"description": "Actor lookup or delivery failed"}}}},
    "/actor": {"get": {"operationId": "localActor", "responses": {"200": {"description": "Development Application actor with ephemeral RSA public key"}}}}
  },
  "components": {"schemas": {
    "NoteKind": {"oneOf": [{"type": "object", "required": ["type"], "properties": {"type": {"const": "post"}}}, {"type": "object", "required": ["type", "target"], "properties": {"type": {"enum": ["reply", "turbo", "turbo_quote"]}, "target": {"type": "string", "format": "uuid"}}}]},
    "CreateNote": {"type": "object", "required": ["id", "author", "content", "kind"], "properties": {"id": {"type": "string", "format": "uuid"}, "author": {"type": "string", "format": "uuid"}, "content": {"type": "string", "description": "Nonblank except turbo, which requires empty text"}, "kind": {"$ref": "#/components/schemas/NoteKind"}}},
    "TimelineEntry": {"type": "object", "required": ["sequence", "note"], "properties": {"sequence": {"type": "integer", "minimum": 1}, "note": {"type": "object", "required": ["created", "reactions"], "properties": {"created": {"$ref": "#/components/schemas/CreateNote"}, "reactions": {"type": "object", "additionalProperties": {"type": "string"}}}}}}
  }}
}